% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roeder_wasserman_weight.R
\name{findc}
\alias{findc}
\title{Compute constant, c, which makes the sum of weights equal to
the number of tests}
\usage{
findc(thet, alpha = 0.05, c0, m)
}
\arguments{
\item{thet}{effect size}

\item{alpha}{significance level}

\item{c0}{square of the tests' critical value}

\item{m}{number of hypothesis tests}
}
\description{
Compute constant, c, which makes the sum of weights equal to
the number of tests
}
\details{
NOne
}
\examples{
theta = seq(0, 3, .5)
c <- sapply(theta, findc, alpha = .05, c0 = 1, m = 100)

}
\references{
Roeder, Kathryn, and Larry Wasserman. "Genome-wide significance
levels and weighted hypothesis testing." Statistical science: a review journal
of the Institute of Mathematical Statistics 24.4 (2009): 398.
\url{www.stat.cmu.edu/~roeder/publications/statsci.pdf}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
