% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roeder_wasserman_weight.R
\name{fun_c}
\alias{fun_c}
\title{Compute sum of normal cummulative values for different constant}
\usage{
fun_c(c, alpha = 0.05, thet, m)
}
\arguments{
\item{c}{a constant value}

\item{alpha}{Significance level}

\item{thet}{effect size}

\item{m}{number of hypothesis tests}
}
\description{
Compute sum of normal cummulative values for the different constant, c.
 This constant c is defined in the paper Roeder and Waserman (2009).
}
\details{
NOne
}
\examples{
cc = seq(.1, 1, .005)
tot <- sapply(cc, fun_c, alpha = .05, thet = .3, m = 100)
}
\references{
Roeder, Kathryn, and Larry Wasserman. "Genome-wide significance
levels and weighted hypothesis testing." Statistical science: a review journal
of the Institute of Mathematical Statistics 24.4 (2009): 398.
\url{www.stat.cmu.edu/~roeder/publications/statsci.pdf}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
