% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_weight_plots.R
\name{prob_weight_plots}
\alias{prob_weight_plots}
\title{Plot ranks' probabilities and the corresponding weights}
\usage{
prob_weight_plots(ey_index, null_index, m, ey, null, prob, weight)
}
\arguments{
\item{ey_index}{Integer vector, column numbers of the null proportions of a
specific effect}

\item{null_index}{Integer vector, column numbers of the covariate-effects of
a specific null proportion}

\item{m}{Integer, total number of tests}

\item{ey}{Numeric, mean covariate-effect size used in \code{ey_index}}

\item{null}{Numeric, proportion of the null tests used in \code{null_index}}

\item{prob}{Numerics data of the ranks probabilities}

\item{weight}{Numeric data of the weights}
}
\value{
Figure of four plots
}
\description{
\code{OPWeight} package proposed a method to compute the ranks
probabilities of the covariates and the corresponding weights given the
test-effect size. This funciton uses the simulation results of the
probabilities and the weights to plot.
}
\details{
To use this function, one needs to load two data sets generated earlier
1) ranksProb_byEffect_m10000.csv, and 2) Weight_byEffect_cont_m10000.csv then
use the function.
}
\examples{
# ranksProb <- read.csv(system.file("simulations/results/ranksProb_byEffect_m10000.csv",
#                                   package = "OPWpaper"), h = TRUE)
# ranksWeight_cont <- read.csv(system.file("simulations/results/Weight_byEffect_cont_m10000.csv",
#                                          package = "OPWpaper"), h = TRUE)
# prob_weight_plots(ey_index = c(7, 17, 27, 37), null_index = c(14, 17, 19),
# m = 10000, ey = 2, null = 50, prob = ranksProb, weight = ranksWeight_cont)

}
\references{
Hasan and Schliekelman (2017)
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
