% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probRel_filterVstest_effect.R
\name{probRel_filterVstest_effect}
\alias{probRel_filterVstest_effect}
\title{Relationship between filter and test effect sizes}
\usage{
probRel_filterVstest_effect(r, rho, H0, ed, m0, m1, n_ey = 100)
}
\arguments{
\item{r}{rank of the test statistics}

\item{rho}{correlation between the filter and test efect sizes}

\item{H0}{determine null or alternative hypotheisis; H0 = 0 for null and H0 = 1 for alternative}

\item{ed}{mean effect size of the test statistics}

\item{m0}{number of the true null hypothesis}

\item{m1}{number of the true alternative hypothesis}

\item{n_ey}{number of fiter-effects to be generated}
}
\value{
\code{prob} probability of the rank given the mean test effect
}
\description{
Compute the relationship between filter and test effect sizes interms of
the probability of the rank of the test given the test effect sizes
}
\details{
Compute the relationship between filter and test effect sizes interms of
the probability of the rank of the tes given the test effect sizes. The weight
identity is based on the test effect size; however, the ranks probability needs
to compute from the filter effect. Therefore, it is expected that there is a
relationship between the filter and test effect sizes.
}
\examples{
ranks = 1:100
prob_test0 <- sapply(ranks, probRel_filterVstest_effect, rho = .8,
                       H0 = 0, ed = 2, m0 = 90, m1 = 10)

prob_test1 <- sapply(ranks, probRel_filterVstest_effect, rho=.8,
                       H0 = 1, ed = 2, m0 = 90, m1 = 10)

prob0 <- sapply(ranks, prob_rank_givenEffect, et = 0, ey = 2,
                                           m0 = 90, m1 = 10)

prob1 <- sapply(ranks, prob_rank_givenEffect, et = 2, ey = 2,
                                           m0 = 90, m1 = 10)

matplot(1:100, cbind(prob_test0, prob_test1, prob0, prob1),
                     type = "l", xlab = "ranks", ylab = "p(rank | effect)")
legend("topright", legend = c("prob_test0", "prob_test1", "prob0", "prob1"),
                       col = 1:4, lty = 1:4, lwd = 2)

}
\seealso{
\code{\link{prob_rank_givenEffect}}
}
\author{
Mohamad S. Hasan and Paul Schliekelman
}
