% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranksProb_byEffect.R
\name{ranksProb_byEffect}
\alias{ranksProb_byEffect}
\title{Ranks probability of the covariate given the test-effect size}
\usage{
ranksProb_byEffect(i, null, m, nrep = 10000, covariateEffectVec)
}
\arguments{
\item{i}{i Integer, i-th effect size of a vector of effects}

\item{null}{Numeric, proportion of the true null hypothesis}

\item{m}{Integer, total number of hypothesis test}

\item{nrep}{Integer, number of replications for the importance sampling}

\item{covariateEffectVec}{A numeric vector of the covariate-effect sizes}
}
\value{
A numeric matrix of the ranks pobabilities in which each column
corresponds to an effect size
}
\description{
Comnpute the ranks probability for different test-effect sizes
}
\details{
This function compute ranks probabilities for the different effect
sizes. It apply the function \code{prob_ranks_givenEffect} from the
\code{OPWeight} package and compute the probabilities.
}
\examples{
# vector of effect sizes
covariateEffectVec <- c(1, 1.5, 2)

# compute ranks probability matrix
ranksProb_byEffect <- sapply(1:length(covariateEffectVec), ranksProb_byEffect,
             null = .9, m = 100, covariateEffectVec = covariateEffectVec)

}
\seealso{
\code{\link{prob_rank_givenEffect}}
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
