% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranksProb_compare.R
\name{ranksProb_compare}
\alias{ranksProb_compare}
\title{Compare rank probabilities}
\usage{
ranksProb_compare(ey, e.one, m0, m1, sampleSize, effectType = c("continuous",
  "binary"))
}
\arguments{
\item{ey}{mean filter effect sizevary}

\item{e.one}{one test effect that will vary across all tests}

\item{m0}{number of true null tests}

\item{m1}{number of true alternative tests}

\item{sampleSize}{total number of sample generated (use sample size at least 100,000)}

\item{effectType}{type of effect sizes; c("continuous", "binary")}
}
\value{
\code{Data} A data frame containing seven columns; ranks and null and
alternative probabilities of the test from the three approaches
}
\description{
\code{OPWeight} package proposed methods to compute the probabilities
of the rank of test given the effect size. This funciton uses the methods to
compare the rank probabilities from three approahes: 1) simulation,
2) exact formula, and 3) normal approximation
}
\details{
\code{OPWeight} package proposed methods to compute the probabilities
of the rank of test given the effect size. This funciton uses the methods to
compare the rank probabilities from three approahes: 1) simulation,
2) exact formula, and 3) normal approximation\cr

rank may generate missing valuse because of the large effcet size,
therefore, to make a matplot equal vectors size are needed. This procedure
will replace the missing value to make equal sized vector
probability of rank of a null test
}
\examples{
# use sample size at least 100,000 for better result
sampleSize = 1000
probData <- ranksProb_compare(ey = 1, e.one = 2, m0 = 90, m1 = 10,
                         sampleSize = sampleSize, effectType = "binary")

# plots------------
colnames(probData) <- c("ranks", "SH0","SH1","EH0","EH1","AH0","AH1")
matplot(probData[, 1], probData[, 2:5], type = "l", lty = 1:6, col =1:6,
lwd = 2, xlab = "ranks", ylab = "P(rank | effect)")
legend("topright", legend = c("SH0","SH1","EH0","EH1","AH0","AH1"),
                lty = 1:6, col =1:6, lwd = 2)

}
\references{
Hasan and Schliekelman (2017)
}
\seealso{
\code{\link{prob_rank_givenEffect_simu}}
\code{\link{prob_rank_givenEffect_exact}}
\code{\link{prob_rank_givenEffect_approx}}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
