% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranksProb_compare_plots.R
\name{ranksProb_compare_plots}
\alias{ranksProb_compare_plots}
\title{Plots to compare ranks probabilities}
\usage{
ranksProb_compare_plots(m0, effectType = c("binary", "continuous"))
}
\arguments{
\item{m0}{number of true null tests}

\item{effectType}{type of effect sizes; c("continuous", "binary")}
}
\value{
\code{Data} A data frame containing seven columns; ranks and null and
alternative probabilities of the test from the three approaches
}
\description{
\code{OPWeight} package proposed methods to compute the probabilities
of the rank of test given the effect size. This funciton uses the simulation results
obtained from the three approahes: 1) simulation, 2) exact formula, and
3) normal approximation to plot the rank probabilities.
}
\details{
To use this function to plot, one needs to load "simu_prob_rank_givenEffect.RDATA"
and apply for different \code{m0}.
}
\examples{
load(system.file("simulations/results", "simu_prob_rank_givenEffect.RDATA",
                             package = "OPWpaper"), envir = environment())
ranksPlots <- ranksProb_compare_plots(m0 = 90, effectType = "continuous")

}
\references{
Hasan and Schliekelman (2017)
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
