% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif_by_mean.R
\name{runif_by_mean}
\alias{runif_by_mean}
\title{Generate uniform random numbers for a fixed mean}
\usage{
runif_by_mean(mean, n)
}
\arguments{
\item{mean}{Numeric, mean of the uniform random numbers}

\item{n}{Integer, number of observations to be generated}
}
\value{
A numeric vector of uniform random numbers
}
\description{
A function to generate uniform random numbers from a fixed mean
}
\details{
Sometime it is necessary to generate random numbers with a prespecified
mean of the uniform distribution. This function will fullfil the goal
}
\examples{
x = runif_by_mean(mean = 3, n = 100)
summary(x)

}
\seealso{
\code{\link{scale}} \code{\link{runif}}
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
