% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_weight.R
\name{rw_weight}
\alias{rw_weight}
\title{Weight from the Roeder and Waserman (2009) paper}
\usage{
rw_weight(testStat, gamma, alpha, group = 5L, tail = 2L)
}
\arguments{
\item{testStat}{test statistics}

\item{gamma}{smooting parameter}

\item{alpha}{Significance level}

\item{group}{number of group}

\item{tail}{one-tailed or two-tailed hypothesis test}
}
\value{
\code{weight} normalized weights of the tests
}
\description{
Compute weight by splitting test statistics. This method is taken
Roeder and Waserman (2009) paper.
}
\details{
NOne
}
\examples{
# generate test statistics
testStat <- rnorm(100000, 2, 1)

# Compute wiehgt
weight = rw_weight(testStat=testStat, gamma=.05, alpha=.05, group=10, tail=2)

# plot the weight
plot(testStat, weight)

}
\references{
Roeder, Kathryn, and Larry Wasserman. "Genome-wide significance
levels and weighted hypothesis testing." Statistical science: a review journal
of the Institute of Mathematical Statistics 24.4 (2009): 398.
\url{www.stat.cmu.edu/~roeder/publications/statsci.pdf}
}
\seealso{
\code{\link{weight_binary}} \code{\link{weight_continuous}}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
