% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_by_block.R
\name{test_by_block}
\alias{test_by_block}
\title{Generate multivariate test statistics}
\usage{
test_by_block(r, eVec, groupSize, Sigma)
}
\arguments{
\item{r}{Integer, number of the test groups}

\item{eVec}{A numeric vector of the means}

\item{groupSize}{Integer, number of test statistics per group}

\item{Sigma}{A numeric matrix correlations}
}
\value{
\code{test} A numeric vector of multivariate test statistics
}
\description{
This function generate multivariate normal test statistics by
block wise.
}
\details{
Generating a large number of tests statistics from the multivariate normal
distribution is computionally very expensive. Therefore, it is someitme convenient
to genrate parts of tests and then concatenate all the tests to obtian the
desired number of the tests statistics. This is called block wise generation
of the test statistics.
}
\examples{
# mean vector
eVec = runif(1000, 0, 2.5)

# correaltion matrix
Sigma <- matrix(.9, 50, 50) + diag(50)*(1 - .1)

# test statistics
testsStat = test_by_block(r = 1, eVec, groupSize = 50,  Sigma)

}
\seealso{
\code{\link{rmvn}}
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
