% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_byEffect_bin.R
\name{weight_byEffect_bin}
\alias{weight_byEffect_bin}
\title{Weight for the different effect sizes in a binary case}
\usage{
weight_byEffect_bin(i, alpha, null, m, tail = 1L, delInterval,
  covariateEffectVec, datByNull)
}
\arguments{
\item{i}{Integer, i-th effect size of a vector of effects}

\item{alpha}{Numeric, a significance level of the hypothesis tests}

\item{null}{Numeric, proportion of true true null tests}

\item{m}{Integer, total number of hypotheis tests}

\item{tail}{Integer (1 or 2), right-tailed or two-tailed hypothesis test.}

\item{delInterval}{Numeric, interval between the \code{delta} values of a
sequence. Note that, \code{delta} is a Langrangian multiplier, necessary to
normalize the weight}

\item{covariateEffectVec}{A numeric vector of covariate-effect sizes}

\item{datByNull}{A numeric matrix of the ranks pobabilities in which each
column corresponds to an effect size}
}
\value{
A numeics matrix of weights in which each column corresponds to an
effect size
}
\description{
Compute the weights from the ranks probability given the test-effect
sizes when the effect sizes are binary.
}
\details{
This function compute the weights when the effect sizes are
binary by applying the ranks probabilities of the different effect sizes.
It applies the function function \code{weight_binary} from the R package
\code{OPWeight} to comute the weights from a probability matirx.
}
\examples{
# vector of effect sizes
covariateEffectVec <- c(1, 1.5, 2)

# compute probability matrix
ranksProb_byEffect <- sapply(1:length(covariateEffectVec), ranksProb_byEffect,
             null = .9, m = 100, covariateEffectVec = covariateEffectVec)

# compute weights
weightByEffect <- sapply(1:length(covariateEffectVec), weight_byEffect_bin,
                   alpha = .05, null = .9, m = 100, delInterval = .01,
                   covariateEffectVec = covariateEffectVec,
                   datByNull = ranksProb_byEffect)

}
\seealso{
\code{\link{ranksProb_byEffect}}
\code{\link{weight_binary}}
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
