% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roeder_wasserman_weight.R
\name{weightSum_by_c}
\alias{weightSum_by_c}
\title{Find sum of weights for c, function of LaGrange multiplier}
\usage{
weightSum_by_c(c, m, gamma = 0.05, alpha = 0.05, group = 5L, tail = 1L,
  effect_hat)
}
\arguments{
\item{c}{a constant function of LaGrange multiplier}

\item{m}{number of hypothesis tests}

\item{gamma}{smooting parameter}

\item{alpha}{Significance level}

\item{group}{number of groups}

\item{tail}{right-tailed or two-tailed hypothesis test. default is right-tailed test}

\item{effect_hat}{estimated effect size}
}
\value{
\code{sum weight} sum of the weights for each C
}
\description{
Compute sum of weights for a given value of c, the function of
LaGrange multiplier
}
\details{
NOne
}
\examples{
cc = seq(-10, 10, .05)
et <- seq(0, 3, .5)
wgtSum_by_c <- sapply(cc, weightSum_by_c, m = 10000, effect_hat = et)

}
\references{
Roeder, Kathryn, and Larry Wasserman. "Genome-wide significance
levels and weighted hypothesis testing." Statistical science: a review journal
of the Institute of Mathematical Statistics 24.4 (2009): 398.
\url{www.stat.cmu.edu/~roeder/publications/statsci.pdf}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
