% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_weight_plots.R
\name{prob_weight_plots}
\alias{prob_weight_plots}
\title{Plot ranks' probabilities and the corresponding weights}
\usage{
prob_weight_plots(ey_index, null_index, m, ey, null, prob, weight)
}
\arguments{
\item{ey_index}{column numbers of the null proportions of a specific effect}

\item{null_index}{column numbers of the effects of specific null proportion}

\item{m}{total number of tests}

\item{ey}{mean filter effect size used in \code{ey_index}}

\item{null}{proportion of the null tests used in \code{null_index}}

\item{prob}{data of the ranks probability}

\item{weight}{data of the weight}
}
\value{
figure of four plots
}
\description{
\code{OPWeight} package proposed methods to compute the probabilities
of the rank of test given the effect size and the corresponding weights.
This funciton uses the simulation results of the probabilities and the weights to plot.
}
\details{
To use this function to plot, one needs to load two data sets generated earlier
1) ranksProb_byEffect_m10000.csv, and 2) Weight_byEffect_cont_m10000.csv then
the function.
}
\examples{
# just an example
# m = 10000
# ranksProb <- read.csv("ranksProb_byEffect_m10000.csv", h=T)
# ranksWeight <- read.csv("Weight_byEffect_cont_m10000.csv",h=T)
# ranksWeight <- t(t(ranksWeight)/colSums(ranksWeight)*m)
# prob_weight_plots(ey_index = c(7, 17, 27, 37), null_index = c(26, 27, 28),
#         m = 10000, ey = 2, null = 50, prob = ranksProb, weight = ranksWeight)

}
\references{
Hasan and Schliekelman (2017)
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
