% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranksProb_byEffect.R
\name{ranksProb_byEffect}
\alias{ranksProb_byEffect}
\title{Probability of rank of test given effect size}
\usage{
ranksProb_byEffect(i, null, m, nrep = 10000, filterEffectVec)
}
\arguments{
\item{i}{i-th effect size}

\item{null}{proportion of the true null hypothesis}

\item{m}{total number of hypothesis test}

\item{nrep}{number of replication for importance sampling}

\item{filterEffectVec}{a vector of filter effect sizes}
}
\value{
a matrix of ranks pobabilities each column corresponds to an effect size
}
\description{
Comnpute the probability of the ranks for different effect sizes
}
\details{
This function compute ranks probabilities for different effect sizes.
It apply the function \code{prob_ranks_givenEffect} from the \code{OPWeight}
package and comute the probabilities.
}
\examples{
# vector of effect sizes
filterEffectVec <- c(1, 1.5, 2)

# compute probability matrix
ranksProb_byEffect <- sapply(1:length(filterEffectVec), ranksProb_byEffect,
             null = .9, m = 100, filterEffectVec = filterEffectVec)

}
\seealso{
\code{\link{prob_rank_givenEffect}}
}
\author{
Mohamad S. Hasan and Paul Schliekelman
}
