% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roeder_wasserman_weight.R
\name{roeder_wasserman_weight}
\alias{roeder_wasserman_weight}
\title{Weight from the Roeder and Wasserman (2009) paper}
\usage{
roeder_wasserman_weight(pvalue, filter, gamma = 0.05, alpha = 0.05,
  group = 5L, tail = 1L)
}
\arguments{
\item{pvalue}{a vector of pvalues of the test statistics}

\item{filter}{a vector of filter statistics}

\item{gamma}{smooting parameter}

\item{alpha}{Significance level}

\item{group}{number of groups}

\item{tail}{right-tailed or two-tailed hypothesis test. default is right-tailed test}
}
\value{
\code{weight} normalized weights of the tests
}
\description{
Compute weights by splitting test statistics raked by the filter statistics.
 This method is taken Roeder and Waserman (2009) paper.
}
\details{
NOne
}
\examples{

#generate pvalues and filter statistics
m = 10000
set.seed(123)
filters = runif(m, min = 0, max = 2.5)          # filter statistics
H = rbinom(m, size = 1, prob = 0.1)             # hypothesis true or false
tests = rnorm(m, mean = H * filters)            # Z-score
pvals = 1 - pnorm(tests)                        # pvalue

# Compute wiehgt
weight = roeder_wasserman_weight(pvalue = pvals, filter = filters)

# plot the weight
plot(weight)

# compute number of rejections
Data <- tibble(tests, pvals, filters)
OD <- Data[order(Data$filters, decreasing = TRUE), ]
alpha = .05
rwd <- sum(OD$pvals <= alpha*weight/m)
bon <- sum(pvals <= alpha/m)

}
\references{
Roeder, Kathryn, and Larry Wasserman. "Genome-wide significance
levels and weighted hypothesis testing." Statistical science: a review journal
of the Institute of Mathematical Statistics 24.4 (2009): 398.
\url{www.stat.cmu.edu/~roeder/publications/statsci.pdf}
}
\seealso{
\code{\link{weight_binary}} \code{\link{weight_continuous}}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
