% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roeder_wasserman_weight.R
\name{roeder_wasserman_weight}
\alias{roeder_wasserman_weight}
\title{Weight from the Roeder and Wasserman (2009) paper}
\usage{
roeder_wasserman_weight(pvalue, gamma = 0.05, alpha = 0.05, group = 5L,
  tail = 1L, c_interval = 0.01)
}
\arguments{
\item{pvalue}{A numeric vector of ordered p-values sorted  by the covariate}

\item{gamma}{Numeric value of the smoothing parameter}

\item{alpha}{Numeric value of the Significance level}

\item{group}{Integer, number of groups}

\item{tail}{Integer (1 or 2), right-tailed or two-tailed hypothesis test.}

\item{c_interval}{A nuumeric vector of a sequence of intervals between the
\code{c}. Note that, \code{c} is a function of Langrangian multiplier,
necessary to normalize the weight}
}
\value{
\code{weight} A numeric vector of the normalized weights
}
\description{
Compute weights by splitting the test statistics raked by the
covariate. This method is taken from Roeder and Waserman (2009).
}
\details{
None
}
\examples{

#generate pvalues and covariate
m = 10000
set.seed(123)
covariates = runif(m, min = 0, max = 2.5)       # covariate
H = rbinom(m, size = 1, prob = 0.1)             # hypothesis true or false
tests = rnorm(m, mean = H * covariates)         # Z-score
pvals = 1 - pnorm(tests)                        # pvalue

# Compute wiehgt
library(tibble)
Data <- tibble(tests, pvals, covariates)
OD <- Data[order(Data$covariates, decreasing = TRUE), ]

weight = roeder_wasserman_weight(pvalue = OD$pvals)

# plot the weight
plot(weight)

# compute number of rejections
alpha = .05
rwd <- sum(OD$pvals <= alpha*weight/m)
bon <- sum(pvals <= alpha/m)

}
\references{
Roeder, Kathryn, and Larry Wasserman. "Genome-wide significance
levels and weighted hypothesis testing." Statistical science: a review journal
of the Institute of Mathematical Statistics 24.4 (2009): 398.
\url{www.stat.cmu.edu/~roeder/publications/statsci.pdf}
}
\seealso{
\code{\link{weight_binary}} \code{\link{weight_continuous}}
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
