% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif_by_mean.R
\name{runif_by_mean}
\alias{runif_by_mean}
\title{Generate uniform random numbers for a fixed mean}
\usage{
runif_by_mean(mean, n)
}
\arguments{
\item{mean}{mean of the random numbers}

\item{n}{number of observations to be generated}
}
\value{
a vector of random numbers
}
\description{
A function to generate uniform random numbers from a fixed mean
}
\details{
Sometime it is necessary to generate random numbers with a prespecified
mean from the uniform distribution. This function will fullfill the goal
}
\examples{
x = runif_by_mean(mean = 3, n = 100)
summary(x)

}
\seealso{
\code{\link{scale}} \code{\link{runif}}
}
\author{
Mohamad S. Hasan and Paul Schliekelman
}
