% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_fwer.R
\name{simu_fwer}
\alias{simu_fwer}
\title{Simulate Family Wise Error Rate (FWER)}
\usage{
simu_fwer(s, m, alphaVec)
}
\arguments{
\item{s}{number of replication in a simulation}

\item{m}{total number of hypothesis test}

\item{alphaVec}{a vector of significance levels}
}
\value{
a matrix of fwer for different methods
}
\description{
This function simulate family wise error rate or test type I error rate
}
\details{
This function generate pvalues times form \code{uniform(0, 1)} then applying
OPWeight method to obtain the Familly Wise Error Rate (FWER)
}
\examples{
alphaVec = seq(.01, .1, .02)
simVal = 1:3  # in actual case use at least simVal = 1000
typeIerror_mat = sapply(simVal, simu_fwer, m = 10000, alphaVec = alphaVec)

}
\references{
Hasan and Schliekelman (2017)
}
\seealso{
\code{\link{qvalue}}
\code{\link{prob_rank_givenEffect}}
\code{\link{weight_binary}}
\code{\link{weight_continuous}}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
