% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_by_block.R
\name{test_by_block}
\alias{test_by_block}
\title{Generate multivariate test statistics by block wise}
\usage{
test_by_block(r, eVec, groupSize, Sigma)
}
\arguments{
\item{r}{number of test groups}

\item{eVec}{a vector of means}

\item{groupSize}{number of test statistics per group}

\item{Sigma}{correlation matrix}
}
\value{
\code{test} a vector of multivariate test statistics
}
\description{
This function generate multivariate normal test statistics by block wise.
}
\details{
Generating large number of tests from multivariate normal distribution
is computionally very slow. Therefore, it is someitme convenient to genrate
parts of tests and then concatenate all tests to obtian the desired number of
tests statistics. This is called block wise generation of the test statistics.
}
\examples{
# mean vector
eVec = runif(1000, 0, 2.5)

# correaltion matrix
Sigma <- matrix(.9, 50, 50) + diag(50)*(1 - .1)

# test statistics
testsStat = test_by_block(r=1, eVec, groupSize = 50,  Sigma)

}
\seealso{
\code{\link{rmvn}}
}
\author{
Mohamad S. Hasan and Paul Schliekelman
}
