% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roeder_wasserman_weight.R
\name{weightSum_by_c}
\alias{weightSum_by_c}
\title{Find sum of weights for c, function of Langragian multiplier}
\usage{
weightSum_by_c(c, m, gamma = 0.05, alpha = 0.05, group = 5L, tail = 1L,
  effect_hat)
}
\arguments{
\item{c}{Numeric, a constant value (function of Langrangian multiplier)}

\item{m}{Integer, number of the hypothesis tests}

\item{gamma}{Numeric value of smoothing parameter}

\item{alpha}{Numeric value of the Significance level}

\item{group}{Integer, number of groups}

\item{tail}{Integer (1 or 2), right-tailed or two-tailed hypothesis test.}

\item{effect_hat}{Numeric, estimated effect sizes}
}
\value{
\code{sum weight} Numeric, sum of the weights for each C
}
\description{
Compute sum of weights for a given value of c, the function of
Langrangian multiplier
}
\details{
None
}
\examples{
cc = seq(-10, 10, .05)
et <- seq(0, 3, .5)
wgtSum_by_c <- sapply(cc, weightSum_by_c, m = 10000, effect_hat = et)

}
\references{
Roeder, Kathryn, and Larry Wasserman. "Genome-wide significance
levels and weighted hypothesis testing." Statistical science: a review journal
of the Institute of Mathematical Statistics 24.4 (2009): 398.
\url{www.stat.cmu.edu/~roeder/publications/statsci.pdf}
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
